Fix file parser source serializer
